test(react-pi): avoid undeclared react dom dependency
