+++ packages/react-pi/src/runtime/usePiRuntime.ts
@@ -35,6 +35,7 @@ import { createPiThreadState, type PiThreadState } from "./threadState";
 import type { PiClient, PiSendMessageInput, PiThreadMetadata } from "../types";
 import { piExtras } from "./piExtras";
 import type { PiRuntimeExtrasInternal, PiRuntimeOptions } from "./runtimeTypes";
+import { invokePiErrorCallback } from "./invokePiErrorCallback";
 
 const EMPTY_THREAD_STATE = createPiThreadState("__pending__");
 const EMPTY_PROJECTED_MESSAGES: readonly ThreadMessageLike[] = [];
@@ -96,20 +97,6 @@ export const NOOP_CONTROLLER: PiThreadControllerLike = {
 
 const NOOP_ON_NEW = () =>
   Promise.reject(new Error("Pi thread is still initializing"));
-
-const invokePiErrorCallback = (
-  onError: PiRuntimeOptions["onError"],
-  error: unknown,
-) => {
-  try {
-    onError?.(error);
-  } catch (callbackError) {
-    console.error(
-      "[assistant-ui/react-pi] onError callback threw an error",
-      callbackError,
-    );
-  }
-};
 
 const buildExtras = (
   controller: PiThreadControllerLike,
